requirements.txt
pandas==1.1.1
PyYAML==5.3.1
scikit-learn==0.23.2
joblib==0.16.0
